Move to separate object - enum
