+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CaseRetrieverTest.java
@@ -23,7 +23,7 @@ import static uk.gov.hmcts.reform.bulkscan.orchestrator.SampleData.SERVICE_TOKEN
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.SampleData.THE_CASE;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.SampleData.USER_ID;
 import static uk.gov.hmcts.reform.bulkscan.orchestrator.SampleData.USER_TOKEN;
-import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.events.EventPublisher.CASE_TYPE_BULK_SCANNED;
+import static uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd.CaseTypeId.CASE_TYPE_BULK_SCANNED;
 
 @RunWith(MockitoJUnitRunner.class)
 public class CaseRetrieverTest {
@@ -34,11 +34,13 @@ public class CaseRetrieverTest {
 
     private CaseRetriever retriever;
 
+    private static final String caseTypeId = CASE_TYPE_BULK_SCANNED.getId();
+
     @Test
     public void should_retrieve_case_successfully() {
         retriever = new CaseRetriever(authenticator, dataApi);
 
-        given(dataApi.readForCaseWorker(USER_TOKEN, SERVICE_TOKEN, USER_ID, JURSIDICTION, CASE_TYPE_BULK_SCANNED, CASE_REF))
+        given(dataApi.readForCaseWorker(USER_TOKEN, SERVICE_TOKEN, USER_ID, JURSIDICTION, caseTypeId, CASE_REF))
             .willReturn(THE_CASE);
         given(authenticator.createForJurisdiction(JURSIDICTION)).willReturn(AUTH_DETAILS);
 
@@ -59,7 +61,7 @@ public class CaseRetrieverTest {
                 .build()
         );
 
-        given(dataApi.readForCaseWorker(USER_TOKEN, SERVICE_TOKEN, USER_ID, JURSIDICTION, CASE_TYPE_BULK_SCANNED, CASE_REF))
+        given(dataApi.readForCaseWorker(USER_TOKEN, SERVICE_TOKEN, USER_ID, JURSIDICTION, caseTypeId, CASE_REF))
             .willThrow(exception);
         given(authenticator.createForJurisdiction(JURSIDICTION)).willReturn(AUTH_DETAILS);
 
@@ -81,7 +83,7 @@ public class CaseRetrieverTest {
                 .build()
         );
 
-        given(dataApi.readForCaseWorker(USER_TOKEN, SERVICE_TOKEN, USER_ID, JURSIDICTION, CASE_TYPE_BULK_SCANNED, CASE_REF))
+        given(dataApi.readForCaseWorker(USER_TOKEN, SERVICE_TOKEN, USER_ID, JURSIDICTION, caseTypeId, CASE_REF))
             .willThrow(exception);
         given(authenticator.createForJurisdiction(JURSIDICTION)).willReturn(AUTH_DETAILS);
